Updates filter_table to display pagination links whenever paginated.
Apparently in newer versions of ActiveRecord with will_paginate,
calling .paginate will sometimes return an ActiveRecord::Relation
instead of the WillPaginate::Collection we're used to. That broke
our pagination links. See this link for extra info:
http://rubydoc.info/gems/will_paginate/3.0.2/file/README.md

Lacking an apparent clear way to check if a set is paginated
or not, we changed the conditional here to see if the set
responds to the total_pages method that will_paginate adds to
the set.
(cherry picked from commit a874242d5e0bba8a132f036b34f17f4003668910)