bug 858782: don't rely on User object hooks to create entities
https://bugzilla.redhat.com/show_bug.cgi?id=858782

Now that we've changed user login to username, the entities migration
is failing -- since it's using the User model save hook to
set entities, but the model now references the new 'username'
field, which only exists after a later migration.

This patch pulls the save hook logic into the migration itself and
uses the attributes as they existed at the time of this migration.