Removes OpenStack and Rackspace support from new installs
They are not supported in 1.1, so we should not put them in our
list of available ProviderTypes.

This change will not remove existing ProviderTypes, and will not
block them from being manually edited. (The two providers remain
in the whitelist in ProvidersController.) However, on new installs,
we will no longer include them in our list.

Resolves https://bugzilla.redhat.com/show_bug.cgi?id=857551

Signed-off-by: Tzu-Mainn Chen <tzumainn@redhat.com>