BZ 837027 - Pool to Cloud Resource Zone fix
We were referencing a non-existent string, causing "Pool" to be
displayed as a missing translation -- thus it didn't get converted
to product names.

The way we are joining these two strings rather than using a new
string is broken but out of scope. I entered #858827 for that.

This resolves https://bugzilla.redhat.com/show_bug.cgi?id=837027

Signed-off-by: Tzu-Mainn Chen <tzumainn@redhat.com>