BZ 857893 - Product string fix - "Cloud Resource Zone Family"
Mechanical translation gone awry! "Pool" got rewritten to "Cloud
Resource Zone", which doesn't work out for "Pool Family".

We may be able to tweak the "Pool" regex to not match "Pool Family",
but that seems like it opens up all sorts of tricky doors about
where "Pool" is used, so it's easier to just use this in the
fixup section.

https://bugzilla.redhat.com/show_bug.cgi?id=857893