Catalogs API #update action
RM3581

Implementation, specs, cucumbers.

In #update action of the controller, the permission check for altering the
catalog's pool association was changed. Previously the check was performed
always, this was causing problems when updating with just partial
representation of the catalog (when pool_id was not specified at all).  Now the
check is performed only if pool_id is present in params.