Consider 401 code as valid when checking provider connection
dc-core returns 401 if no credentials are passed when doing any openstack request,
but conductor can not send credentials when connecting through provider model.
For provider connection test (valid_framework? method) can be 401
response code considered as valid. Bug is tracked here:
https://issues.apache.org/jira/browse/DTACLOUD-330

The connection test is currently not very usable anyway because for most drivers
it just checks if "deltacloud knows about this driver" but not real end-provider
connection is tested.
This can be improved on Conductor side once the following issue is solved on deltacloud side:
https://issues.apache.org/jira/browse/DTACLOUD-307