BZ864527 - aeolus-conductor should install converge-ui into a clean directory
https://bugzilla.redhat.com/show_bug.cgi?id=864527

If the converge-ui git submodule has been checked out, then the spec
will copy the contents of converge-ui-devel overtop of it.  However
there are some files in the submodule that are not included in the
converge-ui-devel rpm, and those files remain.  Now, we'll clean out
the converge-ui directory as shipped in the source tarball before
copying over the contents of converge-ui-devel.