bz864806 removes the empty admin navigation div from permissions/new page
https://bugzilla.redhat.com/show_bug.cgi?id=864806

This issue comes with the problem that permissions/new is used in both monitor and
administer sections and it is quite hard to decide whether to display the administer
nav or not. This patch removes the empty admin_nav div if no administer navigation
is rendered.

In the future we are combining the Monitor and administer sections together, so the
issue will not be relevant any more.
(cherry picked from commit 12463883d9f7ecba394306188dadec9dfb2fc91a)