BZ:867353 Fixed template XPath in Images Controller
The strange XPath in the images controller can only have been there
to support this CDATA attribute that has appeared in our API for
wrapping templates.

The CDATA XML attribute is used to specify data with in a document
that is not to be parsed.  Is is true that our templates are stored
as XML strings in our template model, however, we do not need nor
want a user to know about this.  It is an implementation detail only.

Signed-off-by: Jason Guiditta <jguiditt@redhat.com>
(cherry picked from commit d06e8016102b576239f497a516f5ef4244cefadb)