Password Reset and Username Recovery implementation
This patch adds mailing functionality to conductor. To make emails work properly,
it is necessary to update configuration in src/config/settings.yml (for developement)
or /etc/aeolus-conductor/settings.yml (for production).

It is necessary to set the action_mailer.default_from to valid and existing email address,
to make the email actually sent. Next action_mailer.default_url_options.host and protocol
serve for the links used in email texts. If one of these is set wrong, the links
used in emails (like link to page for changing the password using password reset token)
will not lead to the right url.