API: Standardize linking to resources
We used to have two ways of linking to other resources in incoming XML:

1. <catalog><pool id='1'/></catalog>
2. <catalog><pool_id>1</pool_id></catalog>

For outgoing XML, we consistently use the first format. This commit
adds a before_filter so that we can use the first format also in
incoming XML.

This does not break backwards compatibility, the second format can still
be used.

The filter is used now in CatalogsController, PoolsController and
ProvidersController. It should also be used in
ProviderAccountsController, but it will involve more refactoring and
writing more specs, so this will be done in a separate commit.

Completes RM3922 (with the exception of ProviderAccountsController).