API: Deployments show action (not all information shown yet)
This implements the show action for deployments, so that deployments can
can be linked to and queried for very basic information. There is some
uncertainty about the rest of the fields (e.g. how to represent current
and available deployment states, which date/time formats to use etc.),
so those fields will be added later.

Closes #181.