Fixed realms mapping when finding instance matches
If a frontend realm was mapped to backend realms and all of these
backend realms were unavailable, then match was not found (correct),
but no error message was created.

With this patch the test "shouldn't match frontend realms mapped to
unavailable realms" in spec/models/instance_spec.rb now passes.