API: Duration serializer according to XML Schema
Implements serialization for durations according to
http://www.w3.org/TR/xmlschema-2/#duration

The serializer only uses units upto days, which is permitted by the
standard. It doesn't use years and months, because these units have a
relative length depending on what are they being added/subtracted
to/from, which would be undesirable for our use case.